Update EmailOrUsernameModelBackend
* Support Django 2.1
* Add attribution for the StackOverflow answer it
  was copied from
* Rename the containing module to be more explanatory
* Update comments for clarity
