--- concordia/email_username_backend.py
@@ -1,27 +1,0 @@
-from django.contrib.auth import get_user_model
-from django.contrib.auth.backends import ModelBackend
-from django.db.models import Q
-
-
-class EmailOrUsernameModelBackend(ModelBackend):
-    """
-    ModelBackEnd to allow authentication with either username or email address.
-    """
-
-    def authenticate(self, username=None, password=None, **kwargs):
-        user_model = get_user_model()
-        if username is None:
-            username = kwargs.get(user_model.USERNAME_FIELD)
-        # `username` field does not restring using `@`, so technically email can be
-        # as username and email, even with different users
-        users = user_model._default_manager.filter(
-            Q(**{user_model.USERNAME_FIELD: username}) | Q(email__iexact=username)
-        )
-        # check for any password match
-        for user in users:
-            if user.check_password(password):
-                return user
-        if not users:
-            # Run the default password hasher once to reduce the timing
-            # difference between an existing and a non-existing user (#20760).
-            user_model().set_password(password)
